Moved colors.dart file into styles
all files using colors.dart is autoupdated
