+++ lib/conversational_pages/introduction.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import '../main/colors.dart';
+import '../styles/colors.dart';
 import '../styles/spacing.dart';
 import '../widgets/bottom_nav_2.dart';
