+++ lib/styles/colors.dart
